@@ -4,7 +4,7 @@ version: 0.3.0+2
 homepage: https://github.com/erf/asset_cache
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
