@@ -16,16 +16,16 @@ class AssetCache<T> extends GenericCache<T> {
   AssetDecoder<T> decoder;
 
   /// optional asset bundle
-  AssetBundle bundle;
+  AssetBundle? bundle;
 
   /// add asset decoder in constructor, and optional bundle
-  AssetCache(this.decoder, {this.bundle, String basePath})
+  AssetCache(this.decoder, {this.bundle, String? basePath})
       : super(basePath: basePath);
 
   /// load an asset from a bundle given a key and optional base path and
   /// decode the asset using a given decoder
   Future<T> loadAsset(String name) async {
-    final String key = basePath == null ? name : p.join(basePath, name);
+    final String key = p.join(basePath ?? '', name);
     return await decoder(await (bundle ?? rootBundle).load(key));
   }
 }
