@@ -6,7 +6,7 @@ import '../asset_cache.dart';
 
 /// An asset caches for ByteData
 class ByteDataAssetCache extends AssetCache<ByteData> {
-  ByteDataAssetCache({String basePath, AssetBundle bundle})
+  ByteDataAssetCache({String? basePath, AssetBundle? bundle})
       : super((ByteData data) => Future.value(data),
             basePath: basePath, bundle: bundle);
 }
