@@ -6,7 +6,7 @@ import '../asset_cache.dart';
 
 /// An asset caches for json files
 class JsonAssetCache extends AssetCache<dynamic> {
-  JsonAssetCache({String basePath, AssetBundle bundle})
+  JsonAssetCache({String? basePath, AssetBundle? bundle})
       : super(
             (ByteData data) => Future.value(
                 jsonDecode(utf8.decode(data.buffer.asUint8List()))),
