@@ -6,7 +6,7 @@ import '../asset_cache.dart';
 
 /// An asset caches for strings
 class StringAssetCache extends AssetCache<String> {
-  StringAssetCache({String basePath, AssetBundle bundle})
+  StringAssetCache({String? basePath, AssetBundle? bundle})
       : super(
             (ByteData data) =>
                 Future.value(utf8.decode(data.buffer.asUint8List())),
